Refactor core: split models and rename config
Split and reorganize core model code: model_config.py was renamed to config.py and exported as ModelConfig in core.__init__.py. Introduced compiled_model.py and solved_model.py (moved CompiledModel and SolvedModel implementations out of solver.py) to separate data structures and runtime logic. Updated solver.py to import CompiledModel and SolvedModel and cleaned up long in-file classes; adjusted several function signatures to use explicit None unions for optional args. Updated kalman package imports to reference the new core.config and core.solved_model locations. This change modularizes the codebase and improves separation of concerns between model definitions, compiled artifacts, and solver/runtime behavior.
