+++ SymbolicDSGE/core/__init__.py
@@ -1,4 +1,4 @@
-from .model_config import ModelConfig
+from .config import ModelConfig
 from .model_parser import ModelParser
 from .solver import DSGESolver
 from .shock_generators import Shock
